build: refactor dragondrop package to only access remote NLP Engine
