--- main/internal/implementations/dragon_drop/isolated_test.go
@@ -1,15 +1,0 @@
-package dragondrop
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestNewIsolatedDragonDrop(t *testing.T) {
-	// When
-	dragonDrop := NewIsolatedDragonDrop()
-
-	// Then
-	assert.NotNil(t, dragonDrop)
-}
